requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-background-tasks==1.2.5
django-bootstrap4==2.3.1
django-compat==1.0.15
django-crispy-forms==1.11.1
django-rest-auth==0.9.5
djangorestframework==3.12.2
idna==2.10
newsapi-python==0.2.6
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.4
